[TS Calls] Separate intrinsic model implementations
